Split glib off into its own crate
